pyproject.toml
[project]
name = "security-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]
